First step of refactoring
