+++ hack/runtime-migrator/cmd/backup/main.go
@@ -1,0 +1,44 @@
+package main
+
+import (
+	"fmt"
+	"github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/config"
+	"log/slog"
+	"os"
+	logf "sigs.k8s.io/controller-runtime/pkg/log"
+	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+	"time"
+)
+
+const expirationTime = 60 * time.Minute
+
+func main() {
+	slog.Info("Starting runtime-backuper")
+	cfg := config.NewConfig()
+
+	opts := zap.Options{
+		Development: true,
+	}
+	logger := zap.New(zap.UseFlagOptions(&opts))
+	logf.SetLogger(logger)
+
+	gardenerNamespace := fmt.Sprintf("garden-%s", cfg.GardenerProjectName)
+
+	_, err := config.SetupKubernetesKubeconfigProvider(cfg.GardenerKubeconfigPath, gardenerNamespace, expirationTime)
+	if err != nil {
+		slog.Error(fmt.Sprintf("Failed to create kubeconfig provider: %v", err))
+		os.Exit(1)
+	}
+
+	_, err = config.CreateKcpClient(&cfg)
+	if err != nil {
+		slog.Error("Failed to create kcp client", slog.Any("error", err))
+		os.Exit(1)
+	}
+
+	_, err = config.SetupGardenerShootClient(cfg.GardenerKubeconfigPath, gardenerNamespace)
+	if err != nil {
+		slog.Error("Failed to setup Gardener shoot client", slog.Any("error", err))
+		os.Exit(1)
+	}
+}
