+++ hack/runtime-migrator/input/runtimeids_sample.txt
@@ -1,2 +1,2 @@
 7c6b4bd3-09d3-462f-b27c-992727b62eec
-another-id"
+another-id
