+++ hack/runtime-migrator/internal/config/client.go
@@ -1,0 +1,94 @@
+package config
+
+import (
+	"fmt"
+	"github.com/gardener/gardener/pkg/apis/core/v1beta1"
+	gardener_types "github.com/gardener/gardener/pkg/client/core/clientset/versioned/typed/core/v1beta1"
+	v1 "github.com/kyma-project/infrastructure-manager/api/v1"
+	"github.com/kyma-project/infrastructure-manager/pkg/gardener"
+	"github.com/kyma-project/infrastructure-manager/pkg/gardener/kubeconfig"
+	"github.com/pkg/errors"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/client-go/tools/clientcmd"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"time"
+)
+
+func addToScheme(s *runtime.Scheme) error {
+	for _, add := range []func(s *runtime.Scheme) error{
+		corev1.AddToScheme,
+		v1.AddToScheme,
+	} {
+		if err := add(s); err != nil {
+			return fmt.Errorf("unable to add scheme: %w", err)
+		}
+	}
+	return nil
+}
+
+type GetClient = func() (client.Client, error)
+
+func CreateKcpClient(cfg *Config) (client.Client, error) {
+	restCfg, err := clientcmd.BuildConfigFromFlags("", cfg.KcpKubeconfigPath)
+	if err != nil {
+		return nil, fmt.Errorf("unable to fetch rest config: %w", err)
+	}
+
+	scheme := runtime.NewScheme()
+	if err := addToScheme(scheme); err != nil {
+		return nil, err
+	}
+
+	var k8sClient, _ = client.New(restCfg, client.Options{
+		Scheme: scheme,
+	})
+
+	return k8sClient, nil
+}
+
+func SetupKubernetesKubeconfigProvider(kubeconfigPath string, namespace string, expirationTime time.Duration) (kubeconfig.Provider, error) {
+	restConfig, err := gardener.NewRestConfigFromFile(kubeconfigPath)
+	if err != nil {
+		return kubeconfig.Provider{}, err
+	}
+
+	gardenerClientSet, err := gardener_types.NewForConfig(restConfig)
+	if err != nil {
+		return kubeconfig.Provider{}, err
+	}
+
+	gardenerClient, err := client.New(restConfig, client.Options{})
+	if err != nil {
+		return kubeconfig.Provider{}, err
+	}
+
+	shootClient := gardenerClientSet.Shoots(namespace)
+	dynamicKubeconfigAPI := gardenerClient.SubResource("adminkubeconfig")
+
+	err = v1beta1.AddToScheme(gardenerClient.Scheme())
+	if err != nil {
+		return kubeconfig.Provider{}, errors.Wrap(err, "failed to register Gardener schema")
+	}
+
+	return kubeconfig.NewKubeconfigProvider(shootClient,
+		dynamicKubeconfigAPI,
+		namespace,
+		int64(expirationTime.Seconds())), nil
+}
+
+func SetupGardenerShootClient(kubeconfigPath, gardenerNamespace string) (gardener_types.ShootInterface, error) {
+	restConfig, err := gardener.NewRestConfigFromFile(kubeconfigPath)
+	if err != nil {
+		return nil, err
+	}
+
+	gardenerClientSet, err := gardener_types.NewForConfig(restConfig)
+	if err != nil {
+		return nil, err
+	}
+
+	shootClient := gardenerClientSet.Shoots(gardenerNamespace)
+
+	return shootClient, nil
+}
